# backend/internal/api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Container-scoped firewall control plane. Authentication is bearer-token on TCP listeners and peer-credentials on unix sockets.",
        "title": "Firefik API",
        "contact": {},
        "license": {
            "name": "Apache-2.0"
        },
        "version": "0.3"
    },
    "basePath": "/",
    "paths": {
...
        "/api/audit/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the most recent audit events kept in the in-memory ring buffer. Populated only when `history` is included in `FIREFIK_AUDIT_SINK`. Supports `?limit=\u003cn\u003e` and `?since=\u003cRFC3339\u003e`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Audit history snapshot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "maximum events to return (0 = all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 timestamp; only events strictly after this are returned",
                        "name": "since",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/rules/profiles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the set of built-in rule-set profiles that labels can reference (`firefik.rules.\u003cname\u003e.profile=\u003cprofile\u003e`).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "List rule profiles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.ProfileEntry"
                            }
                        }
                    }
                }
            }
        },
        "/api/rules/templates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns rule templates loaded from `FIREFIK_TEMPLATES_FILE`. Each template has a deterministic 16-hex version derived from its canonical content.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rules"
                ],
                "summary": "List rule templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
...
        "/ws/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "WebSocket endpoint. Frames are JSON `{event,…}` objects (see backend/internal/logstream) and periodic `{event:\"dropped\",count:N}` control messages when a slow consumer lags the hub.",
                "tags": [
                    "logs"
                ],
                "summary": "Subscribe to the live event stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "substring filter applied to each frame before send",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "101": {
                        "description": "switching protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.ProfileEntry": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}